+++ datar/__init__.py
@@ -4,4 +4,4 @@ from .core import operator as _
 from .core import frame_format_patch as _
 from .core.defaults import f
 
-__version__ = '0.2.2'
+__version__ = '0.2.3'